DEL-260: В dto изменил тип price с Int на BigDecimal

See merge request braindelivery1/android!43

+++ feature-route/impl/src/client/kotlin/orderchanges/domain/model/OrderChangedValuesModel.kt
@@ -2,11 +2,12 @@ package orderchanges.domain.model
 
 import cargotype.domain.model.CargoType
 import root.domain.model.extras.OrderExtrasModel
+import java.math.BigDecimal
 
 data class OrderChangedValuesModel(
     val boxes: Int,
     val pallets: Int,
     val cargoType: CargoType,
-    val price: Int,
+    val price: BigDecimal,
     val extras: List<OrderExtrasModel>,
 )
