+++ feature-route/impl/src/client/kotlin/root/domain/RouteRepository.kt
@@ -7,11 +7,12 @@ import root.domain.model.OrderModel
 import root.domain.model.RouteOrderModel
 import root.domain.model.extras.OrderExtrasModel
 import root.presentation.compose.model.RouteStorageModel
+import java.math.BigDecimal
 
 interface RouteRepository {
     suspend fun getOrderDetails(id: Long): RouteOrderModel
     suspend fun getClientOrderDetails(id: Long): OrderModel
-    suspend fun getNewOrderPrice(model: NewOrderModel): Int
+    suspend fun getNewOrderPrice(model: NewOrderModel): BigDecimal
     suspend fun createNewOrder(model: NewOrderModel)
     suspend fun getExtras(): List<OrderExtrasModel>
     suspend fun getStorages(): List<RouteStorageModel>
