+++ feature-route/impl/src/main/kotlin/root/domain/model/extras/OrderExtrasModel.kt
@@ -1,8 +1,10 @@
 package root.domain.model.extras
 
+import java.math.BigDecimal
+
 data class OrderExtrasModel(
     val id: Long,
     val name: String,
-    val price: Int,
+    val price: BigDecimal,
     val priceDescription: OrderPriceDescriptionModel
 )
